move Asset Selector to own file
